third_party/WebKit/LayoutTests/editing/deleting/non-smart-delete.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable> ^foo| bar baz</div>',
        'delete',
        '<div contenteditable>|\u{00A0}bar baz</div>'),
        'Keep space following word should remain');
</script>
